% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enimg.R
\name{base64_img}
\alias{base64_img}
\alias{path_img}
\title{Image to base 64}
\usage{
base64_img(expr)

path_img(expr)
}
\arguments{
\item{expr}{}
}
\value{
a base 64 string
}
\description{
Image to base 64
}
\examples{
if (interactive()){

library(ggplot2)
gg <- ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point()
base64_img_ggplot(gg)

}
}
