% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enjson.R
\name{en_json}
\alias{en_json}
\title{Data to JSON}
\usage{
en_json(expr, ...)
}
\arguments{
\item{expr}{the expression to send to jsonlite::toJSON}
}
\value{

}
\description{
Data to JSON
}
\examples{
en_json(iris)
}
